fix files affected by Repository restructure
