+++ app/src/androidTest/java/com/example/datastore/utils/DataBindingIdlingResourceTest.kt
@@ -16,6 +16,7 @@ import androidx.test.espresso.IdlingResource
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import androidx.test.platform.app.InstrumentationRegistry
 import com.example.datastore.R
+import com.example.datastore.util.mock
 import org.hamcrest.CoreMatchers.`is`
 import org.hamcrest.MatcherAssert.assertThat
 import org.junit.After
