+++ app/src/test-common/java/com/example/datastore/util/MockitoExt.kt
@@ -1,4 +1,4 @@
-package com.example.datastore.utils
+package com.example.datastore.util
 
 import org.mockito.Mockito
 
@@ -7,3 +7,5 @@ import org.mockito.Mockito
  * a kotlin friendly mock that handles generics
  */
 inline fun <reified T> mock(): T = Mockito.mock(T::class.java)
+
+inline fun <reified T> any(): T = Mockito.any(T::class.java)
